+++ calculation/utility/UtilityFunctions.cpp
@@ -13,9 +13,9 @@ using namespace boost::math;
 #include <boost/regex.hpp>
 
 /* returns number of combinations, given 'total' to choose from, choosing 'choose'; where order does not matter and repetition not allowed. */
-double getNumCombinations(size_t total, size_t choose) {
+double getNumCombinations(unsigned int total, unsigned int choose) {
     if (total < choose) return 0.0;
-    return (total == choose) ? 1.0 : factorial<double>(total)/(factorial<double>(choose)*factorial<double>(total - choose));
+    return (total == choose) ? 1.0 : factorial<double>(total)/(factorial<double>(choose) * factorial<double>(total - choose));
 }
 
 //What is the current time? (UTC | Coordinated Universal Time)
